% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{numeric number of trials}

\item{k}{numeric number of successes}
}
\value{
number of combinations successes can occur in trials
}
\description{
Calculates number of combinations in which successes can occur in trials
}
\examples{
x = bin_choose(n = 5, k = 2)

xa = bin_choose(5, 0)

# Multiple success values
xb = bin_choose(5, 1:3)
}
